+++ thriftrw/spec/struct.py
@@ -206,6 +206,13 @@ class StructTypeSpec(TypeSpec):
 
         return self.surface(**kwargs)
 
+    def validate(self, instance):
+        for field in self.fields:
+            field_value = getattr(instance, field.name)
+            if field_value is None:
+                continue
+            field.spec.validate(field_value)
+
     def __str__(self):
         return 'StructTypeSpec(name=%r, fields=%r)' % (self.name, self.fields)